pyproject.toml
[tool.poetry]
name = "image-objects-detector"
version = "0.1.0"
description = ""
authors = ["Grzegorz Szczęsny <grzegorzszczesny794@gmail.com>"]
readme = "README.md"
packages = [{include = "image_objects_detector"}]

[tool.poetry.dependencies]
amqp = "^5.1.1"
boto3 = "^1.28.4"
celery = "^5.3.1"
django = "^4.2.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
django-redis = "^5.3.0"
django-storages = "^1.13.2"
freezegun = "^1.2.2"
psycopg2-binary = "^2.9.6"
numpy = "^1.25.1"
opencv-python = "^4.8.0.74"
python = "^3.11"
redis = "^4.6.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.278"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]

line-length = 140

ignore = ['F403']

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations"
]

target-version = "py311"
